Move process functions to process.rs
