+++ newlib/libc/sys/redox/Makefile.am
@@ -29,7 +29,7 @@ lib_a_DEPENDENCIES = $(extra_objs)
 lib_a_CCASFLAGS = $(AM_CCASFLAGS)
 lib_a_CFLAGS = $(AM_CFLAGS)
 
-$(srcdir)/target/x86_64-unknown-redox/release/libnewlib_redox.a: Cargo.toml src/lib.rs
+$(srcdir)/target/x86_64-unknown-redox/release/libnewlib_redox.a: Cargo.toml src/lib.rs src/macros.rs src/process.rs
 	cd $(srcdir); CC=x86_64-elf-redox-gcc xargo build --target x86_64-unknown-redox --release
 
 # Modified from sys/phoenix
